revised code

+++ src/main/java/com/monitoringsystem/rest_handlers/user_access_mgmt/auth/AuthManager.java
@@ -1,10 +1,8 @@
 package com.monitoringsystem.rest_handlers.user_access_mgmt.auth;
 
-import java.nio.charset.StandardCharsets;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.Base64;
 import java.util.List;
 import java.util.Map;
 
@@ -25,14 +23,14 @@ public class AuthManager
         {
             String sqlQuery = """
                     SELECT auth_token
-                    FROM user_auth_token
+                    FROM user_auth_tokens
                     WHERE email_address = ?
                     """;
             List<Object> sqlParams = List.of(emailAddress);
             ResultSet resultSet = DatabaseOperationsHikari.dbQuery(connection, sqlQuery, sqlParams);
             if (resultSet != null)
             {
-                String jsonString = DatabaseResultsProcessors.processResultsToJson(resultSet);
+                String jsonString = DatabaseResultsProcessors.processResultsToJson(resultSet, connection);
                 Map<String, Map<String, Object>> resultMap = objectMapper.readValue(jsonString, Map.class);
                 storeToken = String.valueOf(resultMap.get("1").get("auth_token"));
             }
@@ -45,6 +43,34 @@ public class AuthManager
 
     }
 
+    public static void insertToken(String emailAddress, String token, String tokenType)
+    {
+        String sqlQuery = "";
+        if ("AUTH_TOKEN".equals(tokenType))
+        {
+            sqlQuery = """
+                    INSERT INTO user_auth_tokens (email_address, auth_token)
+                    VALUES (?, ?)
+                    """;
+        }
+        if ("REG_TOKEN".equals(tokenType))
+        {
+           sqlQuery = """
+                    INSERT INTO user_registration_tokens (email_address, registration_token)
+                    VALUES (?, ?)
+                    """;
+        }
+        try (Connection connection = DatabaseConnectionsHikari.getDbDataSource().getConnection())
+        {
+            List<Object> sqlParams = List.of(emailAddress, token);
+            DatabaseOperationsHikari.dbQuery(connection, sqlQuery, sqlParams);
+        }
+        catch (Exception e)
+        {
+            e.printStackTrace();
+        }
+    }
+
     public static String ephemeralAuthToken(String emailAddress)
     {
         String storedToken = fetchAuthToken(emailAddress);
@@ -55,8 +81,9 @@ public class AuthManager
         else
         {
             String secureToken = TokenGenerator.generateSecureToken();
-            String authToken = "session:" + ":" + secureToken;
-            return Base64.getEncoder().encodeToString(authToken.getBytes(StandardCharsets.UTF_8));
+            String authToken = "session:" + secureToken;
+            insertToken(emailAddress, authToken, "AUTH_TOKEN");
+            return authToken;
         }
     }
 
@@ -70,22 +97,31 @@ public class AuthManager
         else
         {
             String secureToken = TokenGenerator.generateSecureToken();
-            String authToken = "local:" + ":" + secureToken;
-            return Base64.getEncoder().encodeToString(authToken.getBytes(StandardCharsets.UTF_8));
+            String authToken = "local:" + secureToken;
+            insertToken(emailAddress, authToken, "AUTH_TOKEN");
+            return authToken;
         }
     }
 
-    public static void destroyAuthToken(String emailAddress)
+    public static void destroyToken(String emailAddress, String tokenType)
     {
-        Connection connection = null;
-
-        try
+        String sqlQuery = "";
+        if ("AUTH_TOKEN".equals(tokenType))
         {
-            connection = DatabaseConnectionsHikari.getDbDataSource().getConnection();
-            String sqlQuery = """
-            DELETE FROM user_auth_tokens
-            WHERE email_address = ?
-                """;
+            sqlQuery = """
+                    DELETE FROM user_auth_tokens
+                    WHERE email_address = ?
+                    """;
+        }
+        if ("REG_TOKEN".equals(tokenType))
+        {
+           sqlQuery = """
+                    DELETE FROM user_registration_tokens
+                    WHERE email_address = ?
+                    """;
+        }
+        try(Connection connection = DatabaseConnectionsHikari.getDbDataSource().getConnection();)
+        {
             List<Object> sqlParams = List.of(emailAddress);
             DatabaseOperationsHikari.dbQuery(connection, sqlQuery, sqlParams);
         }
@@ -93,55 +129,6 @@ public class AuthManager
         {
             e.printStackTrace();
         }
-        finally
-        {
-            if (connection != null)
-            {
-                try
-                {
-                    connection.close();
-                }
-                catch (SQLException e)
-                {
-                    e.printStackTrace();
-                }
-            }
-        }
-    }
-
-    public static void storeAuthToken(String emailAddress, String authToken)
-    {
-        Connection connection = null;
-
-        try
-        {
-            connection = DatabaseConnectionsHikari.getDbDataSource().getConnection();
-            String sqlQuery = """
-            INSERT INTO user_auth_tokens (email_address, auth_token)
-            VALUES (?, ?)
-                """;
-            List<Object> sqlParams = List.of(emailAddress, authToken);
-            DatabaseOperationsHikari.dbQuery(connection, sqlQuery, sqlParams);
-        }
-        catch (SQLException e)
-        {
-            e.printStackTrace();
-        }
-        finally
-        {
-            if (connection != null)
-            {
-                try
-                {
-                    connection.close();
-                }
-                catch (SQLException e)
-                {
-                    e.printStackTrace();
-                }
-            }
-        }
-
     }
 
 }
